Rewrite some code and adding some validation where it needs

+++ app/Http/Controllers/OrderController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Order;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class OrderController extends Controller
 {
@@ -12,6 +13,14 @@ class OrderController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    protected $rules = [
+        'name'=> 'required|min:3|max:128',
+        'phone' => 'required|min:12|max:15',
+        'description'=>'required|min:3|max:256',
+        'products' => 'required|integer',
+    ];
+
     public function index()
     {
         //
@@ -38,13 +47,18 @@ class OrderController extends Controller
      */
     public function store(Request $request)
     {
+        $validation = Validator::make(Request::all(),$this->rules);
 
-        $order = Order::create(['id'=>$request->input('id'),'status'=>$request->input('status'), 'name'=>$request->input('name'), 'phone'=>$request->input('phone')]);
-        $orderProductIDs = explode(',',$request->input('products'));
-        foreach ($orderProductIDs as $orderProductID) {
-            $order->products()->attach($orderProductID);
+        if ($validation->fails()){
+            return response()->json(['message'=>$validation->errors()],500);
+        } else{
+            $order = Order::create(['id'=>$request->input('id'),'status'=>$request->input('status'), 'name'=>$request->input('name'), 'phone'=>$request->input('phone')]);
+            $orderProductIDs = explode(',',$request->input('products'));
+            foreach ($orderProductIDs as $orderProductID) {
+                $order->products()->attach($orderProductID);
+            }
+            return response()->json(['order'=>$order, 'Ordered Products'=>$order->products()->get()],201);
         }
-        return response()->json(['order'=>$order, 'Ordered Products'=>$order->products()->get()],201);
     }
 
     /**
@@ -81,16 +95,21 @@ class OrderController extends Controller
     public function update(Request $request, int $id)
     {
         //
-        $order = Order::findOrFail($id);
-        $order->products()->detach();
-        $order->name = $request->input('name');
-        $order->phone = $request->input('phone');
-        $orderProductIDs = explode(',',$request->input('products'));
-        foreach ($orderProductIDs as $orderProductID) {
-            $order->products()->attach($orderProductID);
+        $validation = Validator::make(Request::all(),$this->rules);
+        if ($validation->fails()){
+            return response()->json(['message'=>$validation->errors()],500);
+        } else{
+            $order = Order::findOrFail($id);
+            $order->products()->detach();
+            $order->name = $request->input('name');
+            $order->phone = $request->input('phone');
+            $orderProductIDs = explode(',',$request->input('products'));
+            foreach ($orderProductIDs as $orderProductID) {
+                $order->products()->attach($orderProductID);
+            }
+            $order->save();
+            return response()->json(['order'=>$order, 'Ordered Products'=>$order->products()->get()],201);
         }
-        $order->save();
-        return response()->json(['order'=>$order, 'Ordered Products'=>$order->products()->get()],201);
     }
 
     /**
@@ -105,6 +124,6 @@ class OrderController extends Controller
         $order = Order::findOrFail($id);
         $order->products()->detach();
         $order->delete();
-        return  response()->json(null, 201);
+        return  response()->json(null, 204);
     }
 }
